Phase 1 linux port: build chain green
CMake configures and the math-compute core compiles + links into a stub
binary on Linux. GUI/file/CLI layers are excluded for now; replacements
arrive in Phase 2.

See PORTING.md for the full list of adapted files and the planned
phase-2/3 work.

+++ Types.h
@@ -109,7 +109,7 @@ inline HPReal operator""_hp(long double x) {
 }
 
 inline HPReal operator""_hp(unsigned long long x) {
-	return HPReal(x);
+	return HPReal((unsigned long)x);
 }
 
 inline double ToReal(const mpf_class& n) {
